Add copy,/var/log/azure/*/*/* to manifest
Add copy,/var/log/azure/*/*/* to diagnostic manifest as it is commonly needed, so we want it in not just the agents manifest, but also the normal and diagnostic manifest. This will bring parity with the behavior of the Windows manifests, where normal, diagnostic, and agents manifests all collect guest agent logs.